+++ src/main/java/com/nc/med/repo/SupplierRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.nc.med.model.Supplier;
 
-public interface SupplierRepo extends JpaRepository<Supplier, Integer> {
+public interface SupplierRepo extends JpaRepository<Supplier, Long> {
 	Supplier findBySupplierName(String supplierName);
 }
